# backend/services/account-service/api/openapi.yaml
openapi: 3.0.3
info:
  title: Account Service API
  description: Microservice for managing customer accounts and balances.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /accounts:
    post:
      summary: Create a new account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customerId
                - currency
              properties:
                customerId:
                  type: string
                  format: uuid
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Invalid input
    get:
      summary: List accounts for a customer
      operationId: listAccounts
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
  /accounts/{accountId}:
    get:
      summary: Get account details
      operationId: getAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
  /accounts/{accountId}/balance:
    post:
      summary: Update account balance (Deposit/Withdrawal)
      operationId: updateBalance
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - type
                - description
              properties:
                amount:
                  type: number
                  format: double
                type:
                  type: string
                  enum: [DEPOSIT, WITHDRAWAL]
                description:
                  type: string
      responses:
        '204':
          description: Balance updated successfully
        '400':
          description: Invalid input or insufficient funds
        '404':
          description: Account not found

components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        accountNumber:
          type: string
        currency:
          type: string
        balance:
          type: number
          format: double
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []